requirements.txt
twikit>=2.0.0
google-genai>=1.0.0
Pillow>=10.0.0
aiohttp>=3.9.0
aiofiles>=23.0.0
pyyaml>=6.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
